removing the need for connection utility for some base projects
